Day 2
I finished the code. But it is still not working. I got help from Brendon and Paige. After a while we put the code through a text comparer and the codes are almost identical with just comments and some of the names of variables that are wrong. I am glad I have one more day.